Refactor enough to get omise.Charge.list()
